apps/cli/requirements.txt
click>=8.1
httpx>=0.27
PyYAML>=6.0
